Refactor: header -> footer
