+++ src/components/footer/ui/index.ts
