+++ src/components/index.ts
@@ -1,1 +1,1 @@
-export { Footer } from "./Footer";
+export { Footer } from "./footer";
